refactor: modularize large search and gemini service files
- Extracted `search.rs` into focused submodules (helpers, files, content)
- Extracted Gemini caching logic into `GeminiContextCacheManager` and moved schema conversion to `mapper.ts` to simplify `service.ts`
- Applied automated code formatting across the project
