+++ src-tauri/src/services/assistant_service.rs
@@ -151,16 +151,40 @@ mod tests {
 
     #[async_trait]
     impl AssistantRepository for MockAssistantRepository {
-        async fn create_assistant(&self, _id: String, _name: String, _config: String) -> Result<AssistantModel, DbError> { unimplemented!() }
-        async fn get_assistant(&self, _id: &str) -> Result<Option<AssistantModel>, DbError> { unimplemented!() }
-        async fn update_assistant(&self, _id: &str, _name: String, _config: String) -> Result<AssistantModel, DbError> { unimplemented!() }
-        async fn delete_assistant(&self, _id: &str) -> Result<(), DbError> { unimplemented!() }
-        async fn list_assistants(&self) -> Result<Vec<AssistantModel>, DbError> { unimplemented!() }
+        async fn create_assistant(
+            &self,
+            _id: String,
+            _name: String,
+            _config: String,
+        ) -> Result<AssistantModel, DbError> {
+            unimplemented!()
+        }
+        async fn get_assistant(&self, _id: &str) -> Result<Option<AssistantModel>, DbError> {
+            unimplemented!()
+        }
+        async fn update_assistant(
+            &self,
+            _id: &str,
+            _name: String,
+            _config: String,
+        ) -> Result<AssistantModel, DbError> {
+            unimplemented!()
+        }
+        async fn delete_assistant(&self, _id: &str) -> Result<(), DbError> {
+            unimplemented!()
+        }
+        async fn list_assistants(&self) -> Result<Vec<AssistantModel>, DbError> {
+            unimplemented!()
+        }
         async fn search_assistants(&self, _query: &str) -> Result<Vec<AssistantModel>, DbError> {
             Ok(self.assistants.clone())
         }
-        async fn check_assistant_exists(&self, _name: &str) -> Result<bool, DbError> { unimplemented!() }
-        async fn count_assistants(&self) -> Result<u64, DbError> { unimplemented!() }
+        async fn check_assistant_exists(&self, _name: &str) -> Result<bool, DbError> {
+            unimplemented!()
+        }
+        async fn count_assistants(&self) -> Result<u64, DbError> {
+            unimplemented!()
+        }
     }
 
     #[tokio::test]
@@ -170,21 +194,25 @@ mod tests {
                 AssistantModel {
                     id: "1".into(),
                     name: "Rust Expert".into(),
-                    config: r#"{"description":"Helps with Rust","systemPrompt":"You are a Rust dev"}"#.into(),
+                    config:
+                        r#"{"description":"Helps with Rust","systemPrompt":"You are a Rust dev"}"#
+                            .into(),
                     created_at: 0,
                     updated_at: 0,
                 },
                 AssistantModel {
                     id: "2".into(),
                     name: "Python Guru".into(),
-                    config: r#"{"description":"Python scripts","systemPrompt":"You write python"}"#.into(),
+                    config: r#"{"description":"Python scripts","systemPrompt":"You write python"}"#
+                        .into(),
                     created_at: 0,
                     updated_at: 0,
                 },
                 AssistantModel {
                     id: "3".into(),
                     name: "Generic Helper".into(),
-                    config: r#"{"description":"No specific lang","systemPrompt":"Just a helper"}"#.into(),
+                    config: r#"{"description":"No specific lang","systemPrompt":"Just a helper"}"#
+                        .into(),
                     created_at: 0,
                     updated_at: 0,
                 },
@@ -192,22 +220,30 @@ mod tests {
         };
 
         // Match by name
-        let res = AssistantService::search_assistants(&repo, "expert", 10).await.unwrap();
+        let res = AssistantService::search_assistants(&repo, "expert", 10)
+            .await
+            .unwrap();
         assert_eq!(res.len(), 1);
         assert_eq!(res[0].id, "1");
 
         // Match by description
-        let res = AssistantService::search_assistants(&repo, "python scripts", 10).await.unwrap();
+        let res = AssistantService::search_assistants(&repo, "python scripts", 10)
+            .await
+            .unwrap();
         assert_eq!(res.len(), 1);
         assert_eq!(res[0].id, "2");
 
         // Match by systemPrompt
-        let res = AssistantService::search_assistants(&repo, "just a helper", 10).await.unwrap();
+        let res = AssistantService::search_assistants(&repo, "just a helper", 10)
+            .await
+            .unwrap();
         assert_eq!(res.len(), 1);
         assert_eq!(res[0].id, "3");
 
         // Limit
-        let res = AssistantService::search_assistants(&repo, "e", 1).await.unwrap();
+        let res = AssistantService::search_assistants(&repo, "e", 1)
+            .await
+            .unwrap();
         assert_eq!(res.len(), 1);
     }
 }
